Change file structure, Update Readme , add project architecture docs
