--- Login Page/homepage.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="style.css">
-</head>
-<body>
-    <div class="container2">
-        <h1>Welcome, <span id="user-name"></span>!</h1>
-        <p>You have successfully logged in.</p>
-        <button id="logout">Logout</button>
-    </div>
-
-    <script src="home.js" defer></script>
-</body>
-</html>
