+++ README.md
@@ -1,2 +1,18 @@
 # CloudVizz
 CloudVizz , Cloud Cost Visualization made simple
+
+
+## Installation
+
+1. Clone the repository:
+   ```bash
+   git clone https://github.com/Kanishk3813/Mahakumbh_25.git
+2. Change folder
+    ```bash
+    cd /cloudvizz/src
+3. Install dependencies
+    ```bash
+    npm i
+4. Start server 
+    ```bash
+    npm start
