+++ src/database/schema.sql
@@ -1,0 +1,1 @@
+-- null
