+++ src/package.json
@@ -1,0 +1,180 @@
+{
+  "name": "cloudvizz",
+  "version": "1.0.0",
+  "description": "Cloud Cost Visualization made simple",
+  "main": "server.js",
+  "directories": {
+    "test": "tests",
+    "javascript":"javascript"
+  },
+  "dependencies": {
+    
+    "bcrypt": "^5.1.1",
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
+    "express-session": "^1.18.0",
+    "passport": "^0.7.0",
+    "passport-github2": "^0.1.12",
+    "passport-google-oauth20": "^2.0.0",
+    "nodemon": "3.1.3",
+      
+    
+    "abbrev": "^1.1.1",
+    "accepts": "^1.3.8",
+    "agent-base": "^6.0.2",
+    "ansi-regex": "^5.0.1",
+    "aproba": "^2.0.0",
+    "are-we-there-yet": "^2.0.0",
+    "array-flatten": "^1.1.1",
+    "balanced-match": "^1.0.2",
+    "base64url": "^3.0.1",
+    "bcrypt": "^5.1.1",
+    "body-parser": "^1.20.2",
+    "brace-expansion": "^1.1.11",
+    "bytes": "^3.1.2",
+    "call-bind": "^1.0.7",
+    "chownr": "^2.0.0",
+    "color-support": "^1.1.3",
+    "concat-map": "^0.0.1",
+    "console-control-strings": "^1.1.0",
+    "content-disposition": "^0.5.4",
+    "content-type": "^1.0.5",
+    "cookie": "^0.6.0",
+    "cookie-signature": "^1.0.6",
+    "debug": "^2.6.9",
+    "define-data-property": "^1.1.4",
+    "delegates": "^1.0.0",
+    "depd": "^2.0.0",
+    "destroy": "^1.2.0",
+    "detect-libc": "^2.0.3",
+    "dotenv": "^16.4.5",
+    "ee-first": "^1.1.1",
+    "emoji-regex": "^8.0.0",
+    "encodeurl": "^1.0.2",
+    "es-define-property": "^1.0.0",
+    "es-errors": "^1.3.0",
+    "escape-html": "^1.0.3",
+    "etag": "^1.8.1",
+    "express": "^4.19.2",
+    "express-session": "^1.18.0",
+    "finalhandler": "^1.2.0",
+    "forwarded": "^0.2.0",
+    "fresh": "^0.5.2",
+    "fs-minipass": "^2.1.0",
+    "fs.realpath": "^1.0.0",
+    "function-bind": "^1.1.2",
+    "gauge": "^3.0.2",
+    "get-intrinsic": "^1.2.4",
+    "glob": "^7.2.3",
+    "gopd": "^1.0.1",
+    "has-property-descriptors": "^1.0.2",
+    "has-proto": "^1.0.3",
+    "has-symbols": "^1.0.3",
+    "has-unicode": "^2.0.1",
+    "hasown": "^2.0.2",
+    "http-errors": "^2.0.0",
+    "https-proxy-agent": "^5.0.1",
+    "iconv-lite": "^0.4.24",
+    "inflight": "^1.0.6",
+    "inherits": "^2.0.4",
+    "ipaddr.js": "^1.9.1",
+    "is-fullwidth-code-point": "^3.0.0",
+    "make-dir": "^3.1.0",
+    "media-typer": "^0.3.0",
+    "merge-descriptors": "^1.0.1",
+    "methods": "^1.1.2",
+    "mime": "^1.6.0",
+    "mime-db": "^1.52.0",
+    "mime-types": "^2.1.35",
+    "minimatch": "^3.1.2",
+    "minipass": "^5.0.0",
+    "minizlib": "^2.1.2",
+    "mkdirp": "^1.0.4",
+    "ms": "^2.0.0",
+    "negotiator": "^0.6.3",
+    "node-addon-api": "^5.1.0",
+    "node-fetch": "^2.7.0",
+    "nopt": "^5.0.0",
+    "npmlog": "^5.0.1",
+    "oauth": "^0.10.0",
+    "object-assign": "^4.1.1",
+    "object-inspect": "^1.13.1",
+    "on-finished": "^2.4.1",
+    "on-headers": "^1.0.2",
+    "once": "^1.4.0",
+    "parseurl": "^1.3.3",
+    "passport": "^0.7.0",
+    "passport-github2": "^0.1.12",
+    "passport-google-oauth20": "^2.0.0",
+    "passport-oauth2": "^1.8.0",
+    "passport-strategy": "^1.0.0",
+    "path-is-absolute": "^1.0.1",
+    "path-to-regexp": "^0.1.7",
+    "pause": "^0.0.1",
+    "proxy-addr": "^2.0.7",
+    "qs": "^6.11.0",
+    "random-bytes": "^1.0.0",
+    "range-parser": "^1.2.1",
+    "raw-body": "^2.5.2",
+    "readable-stream": "^3.6.2",
+    "rimraf": "^3.0.2",
+    "safe-buffer": "^5.2.1",
+    "safer-buffer": "^2.1.2",
+    "semver": "^7.6.2",
+    "send": "^0.18.0",
+    "serve-static": "^1.15.0",
+    "set-blocking": "^2.0.0",
+    "set-function-length": "^1.2.2",
+    "setprototypeof": "^1.2.0",
+    "side-channel": "^1.0.6",
+    "signal-exit": "^3.0.7",
+    "statuses": "^2.0.1",
+    "string-width": "^4.2.3",
+    "string_decoder": "^1.3.0",
+    "strip-ansi": "^6.0.1",
+    "tar": "^6.2.1",
+    "toidentifier": "^1.0.1",
+    "tr46": "^0.0.3",
+    "type-is": "^1.6.18",
+    "uid-safe": "^2.1.5",
+    "uid2": "^0.0.4",
+    "unpipe": "^1.0.0",
+    "util-deprecate": "^1.0.2",
+    "utils-merge": "^1.0.1",
+    "vary": "^1.1.2",
+    "webidl-conversions": "^3.0.1",
+    "whatwg-url": "^5.0.0",
+    "wide-align": "^1.1.5",
+    "wrappy": "^1.0.2",
+    "yallist": "^4.0.0"
+  },
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "nodemon server.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/CloudVizz/CloudVizz.git"
+  },
+  "keywords": [
+    "aws",
+    "ml",
+    "cloud-computing",
+    "cloud-native",
+    "cost-optimization",
+    "finops",
+    "genai",
+    "cloudcost",
+    "cloudvisualization",
+    "cloudarchitectechture",
+    "costcutting",
+    "architected-view"
+  ],
+  "author": "deepak",
+  "license": "Apache-2.0",
+  "bugs": {
+    "url": "https://github.com/CloudVizz/CloudVizz/issues"
+  },
+  "homepage": "https://github.com/CloudVizz/CloudVizz#readme"
+}
